+++ packages/terra-table/src/subcomponents/Cell.jsx
@@ -136,6 +136,13 @@ const propTypes = {
    * The color to be used for highlighting a column.
    */
   columnHighlightColor: PropTypes.oneOf(Object.values(ColumnHighlightColor)),
+
+  /**
+   * @private
+   * Enables row selection capabilities for the table.
+   * Use 'single' for single row selection and 'multiple' for multi-row selection.
+   */
+  rowSelectionMode: PropTypes.string,
 };
 
 const defaultProps = {
@@ -169,6 +176,7 @@ function Cell(props) {
     firstRowId,
     lastRowId,
     columnHighlightColor,
+    rowSelectionMode,
   } = props;
 
   const cellRef = useRef();
@@ -249,6 +257,9 @@ function Cell(props) {
   }, [isGridContext]);
 
   const handleMouseDown = (event) => {
+    if (rowSelectionMode && (event.button === 2 || hasFocusableElements())) {
+      return;
+    }
     if (!isFocusTrapEnabled) {
       onCellSelect({
         sectionId,
@@ -296,6 +307,9 @@ function Cell(props) {
           break;
         case KeyCode.KEY_SPACE:
           if (onCellSelect) {
+            if (rowSelectionMode && hasFocusableElements()) {
+              return;
+            }
             onCellSelect({
               sectionId,
               rowId,
@@ -391,7 +405,7 @@ function Cell(props) {
   return (
     <CellTag
       id={isRowHeader ? `${tableId}-rowheader-${rowId}` : undefined}
-      ref={isGridContext ? cellRef : undefined}
+      ref={isGridContext || rowSelectionMode ? cellRef : undefined}
       aria-selected={isSelected || undefined}
       aria-label={ariaLabel}
       headers={`${sectionHeaderId}${rowHeaderId}${columnHeaderId}`}