+++ packages/terra-table/CHANGELOG.md
@@ -7,6 +7,9 @@
 * Removed
   * `boundingRef` is no longer necessary for the section header's sticky behavior for tables. This change is still passive in nature as it doesn't change any functionality for consumers.
 
+* Changed
+  * Prevent right click from selecting row in row selection mode.
+
 ## 5.13.0 - (March 14, 2024)
 
 * Added